move searchHandler to separate file
